2nd

+++ functions/src/routes/submissionRouter.ts
@@ -10,14 +10,17 @@ const errorResponse = (error: any, res: any) => {
   res.status(500).json({ message: "Internal Server Error" });
 };
 
-submissionRouter.get("/", async (req, res) => {
+submissionRouter.get("/:uid", async (req, res) => {
   try {
+    const uid: string | undefined = req.params.uid;
     const client = await getClient();
+    const query: any = { uid };
     const results = await client
       .db()
       .collection<Submission>("submissions")
-      .find()
+      .find(query)
       .toArray();
+    res.status(200);
     res.json(results);
   } catch (err) {
     errorResponse(err, res);
@@ -38,14 +41,15 @@ submissionRouter.post("/", async (req, res) => {
   }
 });
 
-submissionRouter.delete("/:id", async (req, res) => {
+submissionRouter.delete("/:id/user/:uid", async (req, res) => {
   try {
+    const uid: string | undefined = req.params.uid;
     const id: string = req.params.id;
     const client = await getClient();
     const results = await client
       .db()
       .collection<Submission>("submissions")
-      .deleteOne({ _id: new ObjectId(id) });
+      .deleteOne({ $and: [{ _id: new ObjectId(id) }, { uid }] });
     if (results.deletedCount) {
       res.sendStatus(204);
     } else {
@@ -57,14 +61,15 @@ submissionRouter.delete("/:id", async (req, res) => {
   }
 });
 
-submissionRouter.get("/:id", async (req, res) => {
+submissionRouter.get("/:id/user/:uid", async (req, res) => {
   try {
+    const uid: string | undefined = req.params.uid;
     const id: string = req.params.id;
     const client = await getClient();
     const result = await client
       .db()
       .collection<Submission>("submissions")
-      .findOne({ _id: new ObjectId(id) });
+      .findOne({ $and: [{ _id: new ObjectId(id) }, { uid }] });
     if (result) {
       res.status(200);
       res.json(result);
